добавлены ссылки для других страниц
