+++ src/components/app/app.tsx
@@ -3,10 +3,10 @@ import { Route, BrowserRouter, Routes } from 'react-router-dom';
 import { HelmetProvider } from 'react-helmet-async';
 import { AppRoute, AuthorizationStatus } from '../../const';
 import MyListPage from '../../pages/my-list-page/my-list-page';
-import SignInPage from '../../pages/sign-in-page/sign-in-page';
-import MoviePage from '../../pages/movie-page/movie-page';
+import SignInPage from '../../pages/sign-in-pages/sign-in-page/sign-in-page';
+import MoviePage from '../../pages/movie-pages/movie-page/movie-page';
 import AddReviewPage from '../../pages/add-review-page/add-review-page';
-import PlayerPage from '../../pages/player-page/player-page';
+import PlayerPage from '../../pages/player-pages/player-page/player-page';
 import NotFoundScreen from '../../pages/not-found-page/not-found-page';
 import PrivateRoute from '../private-route/private-route';
