+++ src/pages/movie-pages/movie-page-in-list/movie-page-in-list.tsx
@@ -118,16 +118,17 @@ function MoviePageInListScreen(): JSX.Element {
                 <p>
                   In the 1930s, the Grand Budapest Hotel is a popular European
                   ski resort, presided over by concierge Gustave H. (Ralph
-                  Fiennes). Zero, a junior lobby boy, becomes Gustave&apos;s friend
-                  and protege.
+                  Fiennes). Zero, a junior lobby boy, becomes Gustave&apos;s
+                  friend and protege.
                 </p>
 
                 <p>
                   Gustave prides himself on providing first-class service to the
-                  hotel&apos;s guests, including satisfying the sexual needs of the
-                  many elderly women who stay there. When one of Gustave&apos;s
-                  lovers dies mysteriously, Gustave finds himself the recipient
-                  of a priceless painting and the chief suspect in her murder.
+                  hotel&apos;s guests, including satisfying the sexual needs of
+                  the many elderly women who stay there. When one of
+                  Gustave&apos;s lovers dies mysteriously, Gustave finds himself
+                  the recipient of a priceless painting and the chief suspect in
+                  her murder.
                 </p>
 
                 <p className="film-card__director">
