+++ src/pages/movie-pages/movie-page-reviews/movie-page-reviews.tsx
@@ -113,8 +113,8 @@ function MoviePageReviewScreen(): JSX.Element {
                       <p className="review__text">
                         Discerning travellers and Wes Anderson fans will
                         luxuriate in the glorious Mittel-European kitsch of one
-                        of the director&apos;s funniest and most exquisitely designed
-                        films in years.
+                        of the director&apos;s funniest and most exquisitely
+                        designed films in years.
                       </p>
 
                       <footer className="review__details">
@@ -132,10 +132,10 @@ function MoviePageReviewScreen(): JSX.Element {
                     <blockquote className="review__quote">
                       <p className="review__text">
                         Anderson&apos;s films are too precious for some, but for
-                        those of us willing to lose ourselves in them, they&apos;re a
-                        delight. &quot;The Grand Budapest Hotel&quot; is no different,
-                        except that he has added a hint of gravitas to the mix,
-                        improving the recipe.
+                        those of us willing to lose ourselves in them,
+                        they&apos;re a delight. &quot;The Grand Budapest
+                        Hotel&quot; is no different, except that he has added a
+                        hint of gravitas to the mix, improving the recipe.
                       </p>
 
                       <footer className="review__details">
@@ -152,9 +152,9 @@ function MoviePageReviewScreen(): JSX.Element {
                   <div className="review">
                     <blockquote className="review__quote">
                       <p className="review__text">
-                        I didn&apos;t find it amusing, and while I can appreciate the
-                        creativity, it&apos;s an hour and 40 minutes I wish I could
-                        take back.
+                        I didn&apos;t find it amusing, and while I can
+                        appreciate the creativity, it&apos;s an hour and 40
+                        minutes I wish I could take back.
                       </p>
 
                       <footer className="review__details">
