+++ src/pages/movie-pages/movie-page/movie-page.tsx
@@ -1,4 +1,7 @@
-function MoviePageDetails(): JSX.Element {
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../../const';
+
+function MoviePageScreen(): JSX.Element {
   return (
     <>
       <section className="film-card film-card--full">
@@ -14,11 +17,11 @@ function MoviePageDetails(): JSX.Element {
 
           <header className="page-header film-card__head">
             <div className="logo">
-              <a href="main.html" className="logo__link">
+              <Link to={AppRoute.Main} className="logo__link">
                 <span className="logo__letter logo__letter--1">W</span>
                 <span className="logo__letter logo__letter--2">T</span>
                 <span className="logo__letter logo__letter--3">W</span>
-              </a>
+              </Link>
             </div>
 
             <ul className="user-block">
@@ -66,9 +69,9 @@ function MoviePageDetails(): JSX.Element {
                   <span>My list</span>
                   <span className="film-card__count">9</span>
                 </button>
-                <a href="add-review.html" className="btn film-card__button">
+                <Link to={AppRoute.AddReview} className="btn film-card__button">
                   Add review
-                </a>
+                </Link>
               </div>
             </div>
           </div>
@@ -88,12 +91,12 @@ function MoviePageDetails(): JSX.Element {
             <div className="film-card__desc">
               <nav className="film-nav film-card__nav">
                 <ul className="film-nav__list">
-                  <li className="film-nav__item">
+                  <li className="film-nav__item film-nav__item--active">
                     <a href="#" className="film-nav__link">
                       Overview
                     </a>
                   </li>
-                  <li className="film-nav__item film-nav__item--active">
+                  <li className="film-nav__item">
                     <a href="#" className="film-nav__link">
                       Details
                     </a>
@@ -106,55 +109,41 @@ function MoviePageDetails(): JSX.Element {
                 </ul>
               </nav>
 
-              <div className="film-card__text film-card__row">
-                <div className="film-card__text-col">
-                  <p className="film-card__details-item">
-                    <strong className="film-card__details-name">
-                      Director
-                    </strong>
-                    <span className="film-card__details-value">
-                      Wes Anderson
-                    </span>
-                  </p>
-                  <p className="film-card__details-item">
-                    <strong className="film-card__details-name">
-                      Starring
-                    </strong>
-                    <span className="film-card__details-value">
-                      Bill Murray, <br />
-                      Edward Norton, <br />
-                      Jude Law, <br />
-                      Willem Dafoe, <br />
-                      Saoirse Ronan, <br />
-                      Tony Revoloru, <br />
-                      Tilda Swinton, <br />
-                      Tom Wilkinson, <br />
-                      Owen Wilkinson, <br />
-                      Adrien Brody, <br />
-                      Ralph Fiennes, <br />
-                      Jeff Goldblum
-                    </span>
-                  </p>
-                </div>
+              <div className="film-rating">
+                <div className="film-rating__score">8,9</div>
+                <p className="film-rating__meta">
+                  <span className="film-rating__level">Very good</span>
+                  <span className="film-rating__count">240 ratings</span>
+                </p>
+              </div>
 
-                <div className="film-card__text-col">
-                  <p className="film-card__details-item">
-                    <strong className="film-card__details-name">
-                      Run Time
-                    </strong>
-                    <span className="film-card__details-value">1h 39m</span>
-                  </p>
-                  <p className="film-card__details-item">
-                    <strong className="film-card__details-name">Genre</strong>
-                    <span className="film-card__details-value">Comedy</span>
-                  </p>
-                  <p className="film-card__details-item">
-                    <strong className="film-card__details-name">
-                      Released
-                    </strong>
-                    <span className="film-card__details-value">2014</span>
-                  </p>
-                </div>
+              <div className="film-card__text">
+                <p>
+                  In the 1930s, the Grand Budapest Hotel is a popular European
+                  ski resort, presided over by concierge Gustave H. (Ralph
+                  Fiennes). Zero, a junior lobby boy, becomes Gustave&apos;s
+                  friend and protege.
+                </p>
+
+                <p>
+                  Gustave prides himself on providing first-class service to the
+                  hotel&apos;s guests, including satisfying the sexual needs of
+                  the many elderly women who stay there. When one of
+                  Gustave&apos;s lovers dies mysteriously, Gustave finds himself
+                  the recipient of a priceless painting and the chief suspect in
+                  her murder.
+                </p>
+
+                <p className="film-card__director">
+                  <strong>Director: Wes Anderson</strong>
+                </p>
+
+                <p className="film-card__starring">
+                  <strong>
+                    Starring: Bill Murray, Edward Norton, Jude Law, Willem Dafoe
+                    and other
+                  </strong>
+                </p>
               </div>
             </div>
           </div>
@@ -176,9 +165,9 @@ function MoviePageDetails(): JSX.Element {
                 />
               </div>
               <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">
+                <Link to={AppRoute.Film} className="small-film-card__link">
                   Fantastic Beasts: The Crimes of Grindelwald
-                </a>
+                </Link>
               </h3>
             </article>
 
@@ -192,9 +181,9 @@ function MoviePageDetails(): JSX.Element {
                 />
               </div>
               <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">
+                <Link to={AppRoute.Film} className="small-film-card__link">
                   Bohemian Rhapsody
-                </a>
+                </Link>
               </h3>
             </article>
 
@@ -208,9 +197,9 @@ function MoviePageDetails(): JSX.Element {
                 />
               </div>
               <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">
+                <Link to={AppRoute.Film} className="small-film-card__link">
                   Macbeth
-                </a>
+                </Link>
               </h3>
             </article>
 
@@ -224,9 +213,9 @@ function MoviePageDetails(): JSX.Element {
                 />
               </div>
               <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">
+                <Link to={AppRoute.Film} className="small-film-card__link">
                   Aviator
-                </a>
+                </Link>
               </h3>
             </article>
           </div>
@@ -234,11 +223,11 @@ function MoviePageDetails(): JSX.Element {
 
         <footer className="page-footer">
           <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
+            <Link to={AppRoute.Main} className="logo__link logo__link--light">
               <span className="logo__letter logo__letter--1">W</span>
               <span className="logo__letter logo__letter--2">T</span>
               <span className="logo__letter logo__letter--3">W</span>
-            </a>
+            </Link>
           </div>
 
           <div className="copyright">
@@ -250,4 +239,4 @@ function MoviePageDetails(): JSX.Element {
   );
 }
 
-export default MoviePageDetails;
+export default MoviePageScreen;
