+++ src/pages/not-found-page/not-found-page.tsx
@@ -1,5 +1,5 @@
-import { Link } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
+import { Link } from 'react-router-dom';
 
 function NotFoundPage(): JSX.Element {
   return (
