+++ src/pages/sign-in-pages/sign-in-error-page/sign-in-error-page.tsx
@@ -1,13 +1,16 @@
-function SignInMessageScreen(): JSX.Element {
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../../const';
+
+function SignInErrorPage(): JSX.Element {
   return (
     <div className="user-page">
       <header className="page-header user-page__head">
         <div className="logo">
-          <a href="main.html" className="logo__link">
+          <Link to={AppRoute.Main} className="logo__link">
             <span className="logo__letter logo__letter--1">W</span>
             <span className="logo__letter logo__letter--2">T</span>
             <span className="logo__letter logo__letter--3">W</span>
-          </a>
+          </Link>
         </div>
 
         <h1 className="page-title user-page__title">Sign in</h1>
@@ -16,13 +19,10 @@ function SignInMessageScreen(): JSX.Element {
       <div className="sign-in user-page__content">
         <form action="#" className="sign-in__form">
           <div className="sign-in__message">
-            <p>
-              We can’t recognize this email <br /> and password combination.
-              Please try again.
-            </p>
+            <p>Please enter a valid email address</p>
           </div>
           <div className="sign-in__fields">
-            <div className="sign-in__field">
+            <div className="sign-in__field sign-in__field--error">
               <input
                 className="sign-in__input"
                 type="email"
@@ -63,11 +63,11 @@ function SignInMessageScreen(): JSX.Element {
 
       <footer className="page-footer">
         <div className="logo">
-          <a href="main.html" className="logo__link logo__link--light">
+          <Link to={AppRoute.Main} className="logo__link">
             <span className="logo__letter logo__letter--1">W</span>
             <span className="logo__letter logo__letter--2">T</span>
             <span className="logo__letter logo__letter--3">W</span>
-          </a>
+          </Link>
         </div>
 
         <div className="copyright">
@@ -78,4 +78,4 @@ function SignInMessageScreen(): JSX.Element {
   );
 }
 
-export default SignInMessageScreen;
+export default SignInErrorPage;
