+++ src/pages/sign-in-pages/sign-in-message-page/sign-in-message-page.tsx
@@ -1,13 +1,16 @@
-function SignInErrorPage(): JSX.Element {
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../../const';
+
+function SignInMessageScreen(): JSX.Element {
   return (
     <div className="user-page">
       <header className="page-header user-page__head">
         <div className="logo">
-          <a href="main.html" className="logo__link">
+          <Link to={AppRoute.Main} className="logo__link">
             <span className="logo__letter logo__letter--1">W</span>
             <span className="logo__letter logo__letter--2">T</span>
             <span className="logo__letter logo__letter--3">W</span>
-          </a>
+          </Link>
         </div>
 
         <h1 className="page-title user-page__title">Sign in</h1>
@@ -16,10 +19,13 @@ function SignInErrorPage(): JSX.Element {
       <div className="sign-in user-page__content">
         <form action="#" className="sign-in__form">
           <div className="sign-in__message">
-            <p>Please enter a valid email address</p>
+            <p>
+              We can’t recognize this email <br /> and password combination.
+              Please try again.
+            </p>
           </div>
           <div className="sign-in__fields">
-            <div className="sign-in__field sign-in__field--error">
+            <div className="sign-in__field">
               <input
                 className="sign-in__input"
                 type="email"
@@ -60,11 +66,11 @@ function SignInErrorPage(): JSX.Element {
 
       <footer className="page-footer">
         <div className="logo">
-          <a href="main.html" className="logo__link logo__link--light">
+          <Link to={AppRoute.Main} className="logo__link logo__link--light">
             <span className="logo__letter logo__letter--1">W</span>
             <span className="logo__letter logo__letter--2">T</span>
             <span className="logo__letter logo__letter--3">W</span>
-          </a>
+          </Link>
         </div>
 
         <div className="copyright">
@@ -75,4 +81,4 @@ function SignInErrorPage(): JSX.Element {
   );
 }
 
-export default SignInErrorPage;
+export default SignInMessageScreen;
